pathfinder/requirements.txt
django
aio-pika
djangorestframework
django-cors-headers
psycopg2
-e git+https://github.com/Alexander35/amqp_handler.git#egg=amqp_handler
